pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ingestor"
version = "0.3.0"
description = "Document ingestion pipeline for Azure AI Search with intelligent chunking and embeddings"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["azure", "ai-search", "document-intelligence", "embeddings", "chunking", "pdf", "document-processing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "azure-ai-documentintelligence>=1.0.0",
    "azure-search-documents>=11.4.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-core>=1.29.0",
    "markitdown>=0.0.1",
    "Pillow>=10.0.0",
    "pymupdf>=1.23.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "tiktoken>=0.5.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
ui = [
    "gradio>=4.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ingestor = "ingestor.cli:cli_main"
ingestor-ui = "ingestor.ui:launch"

[project.urls]
Homepage = "https://github.com/yourusername/ingestor"
Documentation = "https://github.com/yourusername/ingestor/blob/main/docs/INDEX.md"
Repository = "https://github.com/yourusername/ingestor"
Issues = "https://github.com/yourusername/ingestor/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[tool.setuptools.package-data]
ingestor = ["py.typed"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=ingestor --cov-report=html --cov-report=term"
